% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ggplot_theme_type.R
\name{add_ggplot_theme_type}
\alias{add_ggplot_theme_type}
\title{Add the type of the ggplot2 elements (i.e. "rect", "line" or "text")}
\usage{
add_ggplot_theme_type(.data)
}
\arguments{
\item{.data}{Tibble. An object returned by \code{architekter::\link{raw_to_design_tibble}()}}
}
\value{
The same tibble with an extra column with the ggplot2 type (i.e. "rect", "line" or "text").
}
\description{
Add the type of the ggplot2 elements (i.e. "rect", "line" or "text")
}
\examples{
data(toy_raw_file_content)

data_design <- toy_raw_file_content \%>\% 
  raw_to_design_tibble()

data_design \%>\% 
  add_ggplot_theme_type()
}
