% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architekter-package.R
\docType{package}
\name{architekter-package}
\alias{architekter}
\alias{architekter-package}
\title{architekter: Build a {ggplot} Maquette from Figma File}
\description{
Use a Figma file -a free online UI/UX design tool- to build a theme for your graphics created with {ggplot2}. Create a theme, palettes and scale functions easily.
}
\author{
\strong{Maintainer}: Margot Brard \email{margot@thinkr.fr} (\href{https://orcid.org/0000-0001-6754-0659}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
