% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_font_families.R
\name{check_font_families}
\alias{check_font_families}
\title{Determine if the required font families are present on the system}
\usage{
check_font_families(.data)
}
\arguments{
\item{.data}{Tibble. {ggplot2} elements as created by \code{architekter::\link{extract_ggplot_theme}()}}
}
\value{
A tibble with the correct font family. And a message if one of the fonts is missing.
}
\description{
Determine if the required font families are present on the system
}
\examples{
data(toy_raw_file_content)

toy_raw_file_content \%>\% 
  extract_ggplot_theme() \%>\% 
  check_font_families()
}
