% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_theme_fun.R
\name{create_theme_fun}
\alias{create_theme_fun}
\title{Create a theme functions by using the {ggplot2} elements extracted from the Figma file}
\usage{
create_theme_fun(.data)
}
\arguments{
\item{.data}{Tibble. {ggplot2} elements as created by \code{architekter::\link{extract_ggplot_theme}()}}
}
\value{
A {ggplot2} theme function.
}
\description{
Create a theme functions by using the {ggplot2} elements extracted from the Figma file
}
\examples{
data(toy_raw_file_content)

data_theme_elements <- toy_raw_file_content \%>\% 
  extract_ggplot_theme()

my_theme <- create_theme_fun(.data = data_theme_elements)
  
ggplot(data = iris) + 
  aes(x = Sepal.Length, y = Sepal.Width) + 
  geom_point() + 
  my_theme()
}
