% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ggplot_theme_elements.R
\name{determine_ggplot_theme_elements}
\alias{determine_ggplot_theme_elements}
\title{Determine the ggplot2 theme elements.}
\usage{
determine_ggplot_theme_elements(.data)
}
\arguments{
\item{.data}{Tibble. An object returned by \code{swatch::\link{add_ggplot_theme_type}()}}
}
\value{
A tibble with the values of the ggplot2 theme elements.
}
\description{
Determine the ggplot2 theme elements.
}
\examples{
data(toy_raw_file_content)

toy_raw_file_content \%>\% 
  raw_to_design_tibble() \%>\% 
  add_ggplot_theme_type() \%>\% 
  determine_ggplot_theme_elements()
}
