% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ggplot_theme.R
\name{extract_ggplot_theme}
\alias{extract_ggplot_theme}
\title{Extract the {ggplot2} theme elements from a Figma file}
\usage{
extract_ggplot_theme(.data)
}
\arguments{
\item{.data}{List. An object returned by \code{architekter::\link{get_figma_file_content}()}}
}
\value{
A tibble with the {ggplot2} theme elements extracted from the Figma file.
}
\description{
Extract the {ggplot2} theme elements from a Figma file
}
\examples{
data(toy_raw_file_content)

toy_raw_file_content \%>\% 
  extract_ggplot_theme()
}
