% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_content.R
\name{get_file_content}
\alias{get_file_content}
\title{Connect to Figma API and extract content of a file.}
\usage{
get_file_content(file_key, acess_token = Sys.getenv("FIGMA_TOKEN"))
}
\arguments{
\item{file_key}{Character. Key of the Figma file.}

\item{acess_token}{Character. Figma access token.}
}
\value{
A list. Content or the file.
}
\description{
Connect to Figma API and extract content of a file.
}
\details{
The key of your Figma file corresponds to the character
wRqIvMmymzSPuj0sEhnORb in the url
"https://www.figma.com/file/wRqIvMmymzSPuj0sEhnORb/..." (example).
The acess token can be created in your Figma account
(https://www.figma.com/developers/api Section Access tokens)
}
\examples{
file_content <- get_file_content(file_key = "wRqIvMmymzSPuj0sEhnORb",
                                 acess_token = Sys.getenv("FIGMA_TOKEN"))
}
