% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ggplot_element_design.R
\name{get_ggplot_element_design}
\alias{get_ggplot_element_design}
\title{Extract the {ggplot2} design content of a file.}
\usage{
get_ggplot_element_design(name_ggplot_element, file_content)
}
\arguments{
\item{name_ggplot_element}{Character. Name of the {ggplot2} element, as defined in \link[ggplot2]{theme}.}

\item{file_content}{List. Content of the Figma file, as returned by \link[swatch]{get_file_content}.}
}
\value{
List. The {ggplot2} element properties.
}
\description{
Extract the {ggplot2} design content of a file.
}
\examples{
file_content <- get_file_content(file_key = "wRqIvMmymzSPuj0sEhnORb",
                                 acess_token = Sys.getenv("FIGMA_TOKEN"))

get_ggplot_element_design(name_ggplot_element = "panel_background",
                          file_content = file_content)

get_ggplot_element_design(name_ggplot_element = "panel_grid",
                          file_content = file_content)
}
