% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_to_design_tibble.R
\name{raw_to_design_tibble}
\alias{raw_to_design_tibble}
\title{Transform a raw Figma content to a tibble with only the design description of the main components.}
\usage{
raw_to_design_tibble(.data)
}
\arguments{
\item{.data}{List. An object returned by \code{swatch::\link{get_figma_file_content}()}}
}
\value{
A tibble. Design description of the main components.
}
\description{
Transform a raw Figma content to a tibble with only the design description of the main components.
}
\examples{
data(toy_raw_file_content)
toy_raw_file_content \%>\% raw_to_design_tibble()
}
