% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_toy_raw_file_content.R
\docType{data}
\name{toy_raw_file_content}
\alias{toy_raw_file_content}
\title{Toy raw file extraction of a Figma file.}
\format{
A list.
}
\usage{
toy_raw_file_content
}
\description{
A list corresponding to a raw json.
}
\keyword{datasets}
